requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
adalflow
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.15.0
altair>=5.0.0
groq>=0.4.0
sqlparse>=0.4.0
numpy>=1.24.0
requests>=2.31.0
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
chromadb>=0.4.0
prophet>=1.1.0
scikit-learn>=1.3.0
reportlab>=4.0.0
python-docx>=1.1.0
Pillow>=10.0.0
jinja2>=3.1.0
redis==7.1.0
apscheduler>=3.10.0
